src_scanner/requirements.txt
amqp==2.1.4
appdirs==1.4.2
Babel==2.3.4
beautifulsoup4==4.5.3
billiard==3.5.0.2
boto==2.46.1
bs4==0.0.1
-e ./celery-4.0.2
dj-static==0.0.6
Django==1.10.6
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-phonenumber-field==1.1.0
EasyProcess==0.2.3 
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.20.1
gevent==1.2.1
greenlet==0.4.12
gunicorn==19.6.0
honcho==0.5.0
kombu==4.0.2
MySQL-python==1.2.5
packaging==16.8
phonenumberslite==8.3.1
psycopg2==2.7.1
pycurl==7.43.0
pyparsing==2.1.10
pytz==2016.10
redis==2.10.5
requests==2.13
selenium==3.0.2
six==1.10.0
static3==0.7.0
uWSGI==2.0.14
vine==1.1.3
